chore: moved code-processor-plugin
