+++ packages/core/src/generators/stencil/plugins/get-code-processor-plugins.ts
